Added jsdoc
